Moved points listing to CategoryOfDiagrams, since it involves using another diagram (terminal)
