+++ src/main/scala/math/cat/topos/Topos.scala
@@ -9,6 +9,7 @@ trait Topos { self: Category =>
   type Obj
   type Arrow
   // finite limits should exist
+  
   val Ω: Obj
   
 //  def pow: Obj => Obj
