+++ src/test/scala/math/cat/topos/GrothendieckToposTest.scala
@@ -36,20 +36,20 @@ class GrothendieckToposTest extends Test with TestDiagrams {
   "Subobject classifier" should {
     "exist for _0_" in {
       val topos = new CategoryOfDiagrams(_0_)
-      val omega = topos.Ω
-      val points = omega.points
+      import topos._
+      val points = pointsOf(Ω)
       points.size === 1
     }
 
     "exist for _1_" in {
       val topos = new CategoryOfDiagrams(_1_)
-      val omega = topos.Ω
-      val omega0 = omega("0").toList
+      import topos._
+      val points = pointsOf(Ω)
+      val omega0 = Ω("0").toList
       omega0.size === 2
       val omega00::omega01::Nil = omega0
       omega00.asInstanceOf[Diagram]("0").isEmpty must beTrue
       omega01.asInstanceOf[Diagram]("0").isEmpty must beFalse
-      val points = omega.points
       points.size === 2
 
       points.map(_.toString) ===
@@ -58,13 +58,12 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for _2_" in {
       val topos = new CategoryOfDiagrams(_2_)
-      
-      val omega = topos.Ω
-      val omega0 = omega("0")
+      import topos._
+      val points = pointsOf(Ω)
+      val omega0 = Ω("0")
       omega0.size === 3
-      val omega1 = omega("1")
+      val omega1 = Ω("1")
       omega1.size === 2
-      val points = omega.points.toList
       points.size === 3
       points(0).toString ===
         "Point0(0→(0→{}, 1→{}), 1→(0→{}, 1→{}))"
@@ -78,13 +77,12 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for _3_" in {
       val topos = new CategoryOfDiagrams(_3_)
-
-      val omega = topos.Ω
-      val omega0 = omega("0")
+      import topos._
+      val points = pointsOf(Ω)
+      val omega0 = Ω("0")
       omega0.size === 4
-      val omega1 = omega("1")
+      val omega1 = Ω("1")
       omega1.size === 3
-      val points = omega.points.toList
       points.size === 4
       points(0).toString ===
         "Point0(0→(0→{}, 1→{}, 2→{}), 1→(0→{}, 1→{}, 2→{}), 2→(0→{}, 1→{}, 2→{}))"
@@ -101,9 +99,8 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for ParallelPair" in {
       val topos = new CategoryOfDiagrams(ParallelPair)
-
-      val omega = topos.Ω
-      val points = omega.points.toList
+      import topos._
+      val points = pointsOf(Ω)
       points.size === 3  // out of 5 possible candidates, 2 split by a or by b, so they are not points
       points(0).toString ===
         "Point0(0→(0→{}, 1→{}), 1→(0→{}, 1→{}))"
@@ -117,9 +114,8 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for Pullback" in {
       val topos = new CategoryOfDiagrams(Pullback)
-
-      val omega = topos.Ω
-      val points = omega.points.toList
+      import topos._
+      val points = pointsOf(Ω)
       points.size === 5
       points(0).toString ===
         "Point0(a→(a→{}, b→{}, c→{}), b→(a→{}, b→{}, c→{}), c→(a→{}, b→{}, c→{}))"
@@ -139,9 +135,8 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for Pushout" in {
       val topos = new CategoryOfDiagrams(Pushout)
-
-      val omega = topos.Ω
-      val points = omega.points.toList
+      import topos._
+      val points = pointsOf(Ω)
       points.size === 5
       points(0).toString ===
         "Point0(a→(a→{}, b→{}, c→{}), b→(a→{}, b→{}, c→{}), c→(a→{}, b→{}, c→{}))"
@@ -161,9 +156,8 @@ class GrothendieckToposTest extends Test with TestDiagrams {
 
     "exist for Z3" in {
       val topos = new CategoryOfDiagrams(Z3)
-
-      val omega = topos.Ω
-      val points = omega.points.toList
+      import topos._
+      val points = pointsOf(Ω)
       points.size === 2
       points(0).toString === "Point0(0→(0→{}))"
 
@@ -174,8 +168,8 @@ class GrothendieckToposTest extends Test with TestDiagrams {
   "True and False" should {
     "exist for _0_" in {
       val topos = new CategoryOfDiagrams(_0_)
-      val omega = topos.Ω
-      omega.True === omega.False // that's a degenerate topos
+      import topos._
+      Ω.True === Ω.False // that's a degenerate topos
     }
     
     def checkAt(point0: Any)(mappings: (String, set)*): MatchResult[Any] = {
@@ -299,9 +293,11 @@ class GrothendieckToposTest extends Test with TestDiagrams {
       val Δ: DiagramArrow = NaturalTransformation.build("Δ", Ω, ΩxΩ)(
         (x: Ω.d0.Obj) => Ω.d1.arrow(diagonalMap_Ω(domain.obj(x)))).iHope
       val conjunction = Ω.conjunction
-      Ω.True
-
-      failure("wtf")
+      
+      val point = Ω.True.transform(Δ).asInclusion
+      val classifierForTT = classifyingMap(point)
+      val theyAreTheSame = classifierForTT == conjunction
+      theyAreTheSame === true
     }
   }
 }
